Revert "Rename `replicon_tick` module into `tick`"
This reverts commit 081e1b62eaec0046ddc66195a266d0dd1054e0c6.
