+++ src/server/events.rs
@@ -19,7 +19,7 @@ use crate::{
         channels::{RepliconChannel, RepliconChannels},
         common_conditions::*,
         ctx::{ClientReceiveCtx, ServerSendCtx},
-        tick::RepliconTick,
+        replicon_tick::RepliconTick,
         ClientId,
     },
 };
